Added strong name signing to the EventSourceProxy.
Added strong name signing to the EventSourceProxy. Removed the
visibility of internals to the unit test library, and copied the method
required into the unit test. I know copying the method is not DRY, but
it's better than exposing the internals which complicates strong name
signing which I feel is more important.